+++ src/routes/user/[id]/faves/+page.server.ts
@@ -1,8 +1,8 @@
 import { error } from '@sveltejs/kit';
-import { flickr, FlickrError } from '$lib/server/flickr/client';
+import { FlickrError } from '$lib/server/flickr/client';
 import { resolveUserId } from '$lib/server/flickr/users';
+import { getPersonInfo } from '$lib/server/flickr/people';
 import { getUserFaves } from '$lib/server/flickr/favorites';
-import type { PeopleGetInfoResponse } from '$lib/server/flickr/types';
 import type { PageServerLoad } from './$types';
 
 export const load: PageServerLoad = async ({ params }) => {
@@ -16,18 +16,12 @@ export const load: PageServerLoad = async ({ params }) => {
 		throw err;
 	}
 
-	const [info, faves] = await Promise.all([
-		flickr<PeopleGetInfoResponse>({
-			method: 'flickr.people.getInfo',
-			params: { user_id: userId }
-		}),
-		getUserFaves(userId, 1)
-	]);
+	const [user, faves] = await Promise.all([getPersonInfo(userId), getUserFaves(userId, 1)]);
 
 	return {
 		userKey: params.id,
 		userId,
-		user: info.person,
+		user,
 		faves
 	};
 };