+++ src/routes/user/[id]/photostream/+page.server.ts
@@ -1,8 +1,7 @@
 import { error } from '@sveltejs/kit';
-import { flickr, FlickrError } from '$lib/server/flickr/client';
+import { FlickrError } from '$lib/server/flickr/client';
 import { resolveUserId } from '$lib/server/flickr/users';
-import { getUserPhotos } from '$lib/server/flickr/people';
-import type { PeopleGetInfoResponse } from '$lib/server/flickr/types';
+import { getPersonInfo, getUserPhotos } from '$lib/server/flickr/people';
 import type { PageServerLoad } from './$types';
 
 export const load: PageServerLoad = async ({ params }) => {
@@ -16,18 +15,12 @@ export const load: PageServerLoad = async ({ params }) => {
 		throw err;
 	}
 
-	const [info, photos] = await Promise.all([
-		flickr<PeopleGetInfoResponse>({
-			method: 'flickr.people.getInfo',
-			params: { user_id: userId }
-		}),
-		getUserPhotos(userId, 1)
-	]);
+	const [user, photos] = await Promise.all([getPersonInfo(userId), getUserPhotos(userId, 1)]);
 
 	return {
 		userKey: params.id,
 		userId,
-		user: info.person,
+		user,
 		photos
 	};
 };